Migrate persisted state to the graph store
