+++ .github/workflows/temp-branch-build-and-push-upgrade-hawk.yaml
@@ -3,7 +3,7 @@ name: Branch - Hawk Upgrade Build and push docker image
 on:
   push:
     branches:
-      - "add-snapshot-aurora"
+      - "fix-persisted-state"
 
 concurrency:
   group: '${{ github.workflow }} @ ${{ github.event.pull_request.head.label || github.head_ref || github.ref }}'
